Move browser configuration to correct place
