+++ wdio.conf.e2e.js
@@ -1,14 +1,7 @@
-import * as path from 'node:path';
 import * as fs from 'node:fs';
-import { fileURLToPath } from 'url';
 
 import { config as base } from './wdio.conf.base.js';
 import WebpackDevServerService from './test/support/wdio-webpack-dev-server.js';
-
-// shim __dirname since we're in an ES module
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-global.downloadDir = path.join(__dirname, 'tempDownload');
 
 export const config = {
   ...base,
@@ -21,37 +14,16 @@ export const config = {
     ...base.services,
     [WebpackDevServerService, {}]
   ],
-  // chrome configuration for download testing
-  capabilities: [{
-    browserName: 'chrome',
-    'goog:chromeOptions': {
-      prefs: {
-        'directory_upgrade': true,
-        'prompt_for_download': false,
-        'download.default_directory': downloadDir
-      },
-      args: [
-        'headless',
-        'disable-gpu'
-      ]
-    }
-  },
-  {
-    browserName: 'firefox'
-  },
-  {
-    browserName: 'safari'
-  }],
   onPrepare: function (config, capabilities) {
     // make sure download directory exists
-    if (!fs.existsSync(downloadDir)){
+    if (!fs.existsSync(global.downloadDir)){
         // if it doesn't exist, create it
-        fs.mkdirSync(downloadDir);
+        fs.mkdirSync(global.downloadDir);
     }
   },
   onComplete: function() {
-    console.log(downloadDir)
-    fs.rmSync(downloadDir, {
+    console.log(global.downloadDir)
+    fs.rmSync(global.downloadDir, {
       recursive: true,
       force: true
     });
